borrow_book function RAW
I made the borrow_book function but i have not tested it. So not merged the program with main program.
Do the necessary changes where needed.
register_book.txt is another database which stores the entries of books and person taking the book with the issuing date and returning date.